requirements_0301.txt
accelerate==0.27.2
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
auto_gptq==0.7.0
backoff==2.2.1
bcrypt==4.1.2
bitsandbytes==0.41.1
Brotli==1.0.9
build==1.0.3
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.17.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.0
ffmpy==0.3.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.10.0
gekko==1.0.6
google-auth==2.28.0
googleapis-common-protos==1.62.0
gradio==4.19.1
gradio_client==0.10.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface_hub==0.20.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.4.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.17.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
optimum==1.17.1
orjson==3.9.10
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
peft==0.8.2
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pluggy==1.4.0
posthog==3.4.1
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.23.23
PyMuPDFb==1.23.22
pyparsing==3.1.1
pypdf==4.0.2
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.0.0
pytest-subtests==0.11.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.0
rouge==1.0.1
rpds-py==0.18.0
rsa==4.9
ruff==0.2.2
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
setuptools==69.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==1.26.13
uvicorn==0.27.1
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==10.4
wheel==0.42.0
win-inet-pton==1.1.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
